+++ app/src/main/java/com/example/track/domain/models/abstractLayer/User.kt
@@ -10,6 +10,6 @@ data class User(
     var username: String = NAME_DEFAULT,
     var needsLogin: Boolean = NEEDS_LOGIN,
     var loginCount: Int = LOGIN_COUNT_DEFAULT,
-    var budget: Int = BUDGET_DEFAULT,
+    var budget: Float = BUDGET_DEFAULT,
     var currency: String = CURRENCY_DEFAULT.ticker
 )
